+++ src/main/java/com/fourfourfour/damoa/api/user/dao/JdbcUserDaoImpl.java
@@ -1,8 +1,7 @@
-package com.fourfourfour.damoa.model.dao.user;
+package com.fourfourfour.damoa.api.user.dao;
 
-import com.fourfourfour.damoa.model.dto.user.UserDto;
+import com.fourfourfour.damoa.api.user.dto.UserDto;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
