+++ src/main/java/com/fourfourfour/damoa/common/dto/response/ErrorSerializer.java
@@ -1,4 +1,4 @@
-package com.fourfourfour.damoa.model.dto.response;
+package com.fourfourfour.damoa.common.dto.response;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
