+++ src/main/java/com/fourfourfour/damoa/config/auth/PrincipalDetails.java
@@ -1,6 +1,6 @@
 package com.fourfourfour.damoa.config.auth;
 
-import com.fourfourfour.damoa.model.dto.user.UserDto;
+import com.fourfourfour.damoa.api.user.dto.UserDto;
 import lombok.Data;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
