+++ src/main/java/com/fourfourfour/damoa/config/jwt/JwtAuthorizationFilter.java
@@ -3,8 +3,8 @@ package com.fourfourfour.damoa.config.jwt;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.fourfourfour.damoa.config.auth.PrincipalDetails;
-import com.fourfourfour.damoa.model.dao.user.UserDao;
-import com.fourfourfour.damoa.model.dto.user.UserDto;
+import com.fourfourfour.damoa.api.user.dao.UserDao;
+import com.fourfourfour.damoa.api.user.dto.UserDto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.authentication.AuthenticationManager;
