+++ src/test/java/com/fourfourfour/damoa/model/dao/user/JdbcUserDaoImplTest.java
@@ -1,6 +1,7 @@
 package com.fourfourfour.damoa.model.dao.user;
 
-import com.fourfourfour.damoa.model.dto.user.UserDto;
+import com.fourfourfour.damoa.api.user.dao.UserDao;
+import com.fourfourfour.damoa.api.user.dto.UserDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
